// src/app/shared/model/userInfo.ts
export class UserInfo {
    Name: string;
    Email: string;
    Number: string;
    Block: string;
    Street: string;
    Building: string;
    Floor: string;
    CarModel: string;
    CarColor: string;
    CarNumber: string;

    constructor() {
        this.Name = '';
        this.Email = '';
        this.Number = '';
        this.Block = '';
        this.Street = '';
        this.Building = '';
        this.Floor = '';
        this.CarModel = '';
        this.CarColor = '';
        this.CarNumber = '';
    }
}