// src/app/shared/model/item.ts
export class Item {
  guid: string;
  date?: Date;
  name: string;
  note: string;
  group_Guid: string;
  balance: number;
  avg_Price: number;
  unit1: string;
  qty_Unit1: number;
  Barcode_1: string;
  latin_Name: string;
  company: string;
  provenance: string;
  quality: string;
  model: string;
  dim: string;
  pos: string;
  guid_User: string;
  guid_Currency: string;
  endUser: number;
  vendor: number;
  export: number;
  retail: number;
  lastPrice: number;
  orderLimit: number;
  max_Qty: number;
  min_Qty: number;
  type_Iteme: string;
  vat: number;
  is_Remove: string;
  type: string;
  status: string;
  is_Sales: string;
  image: string;
  sort: number;
  quantity: number;
  id: number;
  extraItems: Item[];

  constructor() {
    this.guid = '';
    this.date = null;
    this.name = '';
    this.note = '';
    this.group_Guid = '';
    this.balance = null;
    this.avg_Price = null;
    this.unit1 = '';
    this.qty_Unit1 = null;
    this.Barcode_1 = '';
    this.latin_Name = '';
    this.company = '';
    this.provenance = '';
    this.quality = '';
    this.model = '';
    this.dim = '';
    this.pos = '';
    this.guid_User = '';
    this.guid_Currency = '';
    this.endUser = null;
    this.vendor = null;
    this.export = null;
    this.retail = null;
    this.lastPrice = null;
    this.orderLimit = null;
    this.max_Qty = null;
    this.min_Qty = null;
    this.type_Iteme = '';
    this.vat = null;
    this.is_Remove = '';
    this.type = '';
    this.status = '';
    this.is_Sales = '';
    this.image = '';
    this.sort = null;
    this.quantity = 0;
    this.id = 0;
    this.extraItems = [];
  }
}